% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatBreak.R
\docType{package}
\name{StatBreak}
\alias{StatBreak}
\title{StatBreak package}
\description{
Through this package, researchers can conduct easy and insightful robustness tests.
   Generally, the provided function 'break' answers the question: How few and which cases would need to be deleted
   for a substantial change in the sample statistic?
   The underlying algorithm searches the space of possible subsamples through genetic evolution.
   Most code for implementing the generic genetic algorithm code was copied and modified from the
   general purpose package 'genalg' by Egon Willighagen and Michel Ballings (currently not maintained).
}
\author{
Hannes Rosenbusch \email{h.rosenbusch@uvt.nl}
}
